Add update function to document db interface and return ids for conversations instead of model instance

+++ streamlitapp/tests/conftest.py
@@ -5,7 +5,8 @@ from sqlalchemy import create_engine, MetaData
 from sqlalchemy.orm import sessionmaker, Session
 from perry.db.models import Base
 from perry.db.operations.agents import create_agent
-from perry.db.operations.conversations import create_conversation
+from perry.db.operations.conversations import create_conversation, update_conversation
+from perry.db.operations.documents import create_document
 from perry.db.operations.users import create_user
 from perry.db.api import connect_agent_to_conversation
 
@@ -34,35 +35,51 @@ def add_agent_to_db(test_db) -> int:
     """Add an agent to the database and return its ID."""
     return create_agent(test_db)
 
-
 @pytest.fixture(scope="function")
 def add_conversation_to_db(test_db) -> int:
     """Add a conversation to the database and return its ID."""
-    return create_conversation(test_db)
+    def _add_conversation_to_db():
+        return create_conversation(test_db)
+    return _add_conversation_to_db
 
+@pytest.fixture(scope="function")
+def create_user_in_db(test_db):
+    """Add a user to the database and return its ID."""
+    def _create_user_in_db(username: str, password: str):
+        return create_user(test_db, username, password)
+
+    return _create_user_in_db
+
+@pytest.fixture(scope="function")
+def add_document_to_db(test_db):
+    """Add a document to the database and return its ID."""
+    def _add_document_to_db():
+        return create_document(test_db)
+    return _add_document_to_db
 
 @pytest.fixture(scope="function")
 def create_connected_agent_conversation_in_db(
     test_db, add_agent_to_db, add_conversation_to_db
 ) -> tuple[int, int]:
     """Add an agent and a conversation to the database, connect them and return their IDs."""
-    connect_agent_to_conversation(test_db, add_agent_to_db, add_conversation_to_db)
+    connect_agent_to_conversation(test_db, add_agent_to_db(), add_conversation_to_db())
     return add_agent_to_db, add_conversation_to_db
 
-
-@pytest.fixture(scope="function")
-def create_user_in_db():
-    """Add a user to the database and return its ID."""
-
-    def _create_user_in_db(test_db, username: str, password: str):
-        return create_user(test_db, username, password)
-
-    return _create_user_in_db
-
-
 # @pytest.fixture(scope="function")
-# def create_connected_agent_conversation_in_db_with_docs():
-#     pass
+# @pytest.mark.parametrize("temp_files", [
+#     [{'name': 'file1', 'contents': 'file1 contents', 'suffix': '.pdf'}],
+#     [{'name': 'file2', 'contents': 'file2 contents', 'suffix': '.pdf'}],
+#     [{'name': 'file3', 'contents': 'file3 contents', 'suffix': '.pdf'}],
+# ], indirect=True)
+# def create_connected_agent_conversation_in_db_with_docs(
+#     create_connected_agent_conversation_in_db, create_user_in_db, temp_files, test_db
+# ):
+#     agent_id, conversation_id = create_connected_agent_conversation_in_db
+#     user_id = create_user_in_db(test_db, 'test_user', 'test_password')
+#     for file_path in temp_files:
+#         doc_id = create_document(test_db)
+#         doc_id = update_document()
+
 
 
 @pytest.fixture(scope="function")
@@ -89,12 +106,14 @@ def temp_files(request):
         with open(path, 'w') as f:
             f.write(contents)
 
-        temp_file_details.append({'path': path, 'name': file_name})
+        temp_file_details.append({'path': path, 'name': file_name, 'suffix': suffix})
     
     yield temp_file_details
 
     for file_detail in temp_file_details:
-        os.remove(file_detail['path'])
+        path = file_detail['path']
+        if os.path.exists(path):
+            os.remove(file_detail['path'])
 
 
 def clear_db(db_session: Session):
